warning as errors options + some additionnal warnings + some fixes

+++ src/importer/r3d_importer_texture.c
@@ -120,7 +120,7 @@ static texture_key_t make_key_texture_job(const texture_job_t* job)
 {
     uint64_t hash = R3D_HASH_FNV_OFFSET_BASIS_64;
 
-    for (int i = 0; i < R3D_ARRAY_SIZE(job->paths); i++)
+    for (int i = 0; i < (int)R3D_ARRAY_SIZE(job->paths); i++)
     {
         if (job->paths[i][0] != '\0')
         {
